BackEnd/requirements.txt
amqp==5.1.1
apturl==0.5.2
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
celery==5.2.7
certifi==2019.11.28
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.4.0
distro-info===0.23ubuntu1
dj-rest-auth==2.2.5
Django==4.1.3
django-admin==2.0.1
django-allauth==0.22.0
django-braces==1.15.0
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==22.1
django-oauth-toolkit==1.2.0
django-oauth-toolkit-fork==0.11.0
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-restframework==0.0.1
django-separatedvaluesfield==0.6.0
django-six==1.0.4
django-templated-mail==1.1.1
django-timezone-field==5.0
django-zoom-meetings==0.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kazam==1.4.5
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
msgpack==1.0.5
netifaces==0.10.4
oauth2-provider==0.0
oauthlib==3.2.2
olefile==0.46
paramiko==2.6.0
pbr==5.11.0
pexpect==4.6.0
Pillow==7.0.0
prompt-toolkit==3.0.38
protobuf==3.6.1
psycopg2-binary==2.9.5
pycairo==1.16.2
pycodestyle==2.10.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.1
python-crontab==2.7.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
redis==4.5.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
screen==1.0.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
tomli==2.0.1
tzdata==2022.7
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.6
wrapt==1.14.1
xkit==0.0.0
xlwt==1.3.0
